Add folder and his files

+++ 0x02-python-import_modules/calculator_1.py
@@ -35,7 +35,7 @@ def mul(a, b):
     Returns:
         The return value. a * b
     """
-    return (a * b)
+    return int(a * b)
 
 
 def div(a, b):
